Python/requirements.txt
paho-mqtt
pyserial
psycopg2-binary
